Fix C4244 convertion warnings
See [Visual Studio Compiler Warning C4244](https://docs.microsoft.com/en-us/cpp/error-messages/compiler-warnings/compiler-warning-levels-3-and-4-c4244).

+++ matplotlibcpp.h
@@ -1817,7 +1817,7 @@ template<typename Numeric>
 bool plot(const std::vector<Numeric>& y, const std::string& format = "")
 {
     std::vector<Numeric> x(y.size());
-    for(size_t i=0; i<x.size(); ++i) x.at(i) = i;
+    for(size_t i = 0; i < x.size(); ++i) x.at(i) = static_cast<Numeric>(i);
     return plot(x,y,format);
 }
 
@@ -1825,7 +1825,7 @@ template<typename Numeric>
 bool plot(const std::vector<Numeric>& y, const std::map<std::string, std::string>& keywords)
 {
     std::vector<Numeric> x(y.size());
-    for(size_t i=0; i<x.size(); ++i) x.at(i) = i;
+    for(size_t i = 0; i < x.size(); ++i) x.at(i) = static_cast<Numeric>(i);
     return plot(x,y,keywords);
 }
 
@@ -1833,7 +1833,7 @@ template<typename Numeric>
 bool stem(const std::vector<Numeric>& y, const std::string& format = "")
 {
     std::vector<Numeric> x(y.size());
-    for (size_t i = 0; i < x.size(); ++i) x.at(i) = i;
+    for(size_t i = 0; i < x.size(); ++i) x.at(i) = static_cast<Numeric>(i);
     return stem(x, y, format);
 }
